requirements.txt
streamlit>=1.34.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.0
chromadb>=0.5.0
pypdf>=4.2.0
sentence-transformers>=2.7.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pandas>=2.2.2
numpy>=1.26.4
